drop sleep() and increase retry/timeout counts for our acceptance tests.
After the vagrant reload, our boxes take a few minutes to recover and start all required services. This is mostly as they're somehow dependent on each other. For example, everything depends on tinc-core. Zookeeper requires a tinc ip, Mesos masters/slaves require zookeeper to be up. Marathon requires mesos
master to be elected as well as a zookeeper connection.
There are monit checks that restart each service if they're not available, although those checks take a few minutes to act.

The changes in this commit, will move the sleep we usually have between a reload of the VMs and the run of the acceptance_tests to simply re-run the acceptance tests up to 'n' times with a longer wait period between them.
this generates more errors in the build.log but can save a couple of minutes vs a static sleep() block.